+++ src/Webiny/Component/Rest/CHANGELOG.md
@@ -7,6 +7,7 @@ CHANGELOG
 * cache now supports multiple drivers
 * removed cache file path from debug headers
 * rest annotation can now be passed from parent class to a child class
+* added resource naming option (`@rest.url`)
 
 1.1
 ---------